corrected datatype of timeOffset to make smooth movement in orbits

+++ modules/space/translation/keplertranslation.cpp
@@ -304,7 +304,7 @@ glm::dvec3 KeplerTranslation::position(const UpdateData& data) const {
     return _orbitPlaneRotation * p;
 }
 // !!! is only used in module/space/rendering/renderablesatellites 
-glm::dvec3 KeplerTranslation::debrisPos(const float& time) const {
+glm::dvec3 KeplerTranslation::debrisPos(const double& time) const {
     if (_orbitPlaneDirty) {
         computeOrbitPlane();
         _orbitPlaneDirty = false;
